requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.9
python-dotenv==1.1.1
langchain==0.3.27
langchain-text-splitters==0.3.9
langchain-core==0.3.72
google-generativeai==0.8.5
pinecone-client==6.0.0
pypdf==4.2.0
python-docx==1.1.2
pydantic==2.11.7
requests==2.32.4
python-dateutil==2.9.0 